*Se implementa Spring Security 6

+++ auth-server/src/main/java/com/fabianbah/auth_server/repositories/CustomerRepository.java
@@ -1,12 +1,11 @@
 package com.fabianbah.auth_server.repositories;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fabianbah.auth_server.entities.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, BigInteger>{
+public interface CustomerRepository extends JpaRepository<Customer, Long>{
     Optional<Customer> findByEmail(String email);
 }
